requirements.txt
python-telegram-bot==20.7
Flask==3.0.3
SQLAlchemy==2.0.36
APScheduler==3.10.4
python-dotenv==1.0.1
passlib==1.7.4
